backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
anthropic==0.89.0
openai==2.30.0
google-genai==1.73.1
pydantic==2.12.5
python-dotenv==1.2.2
SQLAlchemy==2.0.27
aiosqlite==0.22.1
reportlab==4.4.10
PyYAML==6.0.3
python-multipart==0.0.24
httpx==0.28.1
aiofiles==25.1.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-mock==3.15.1
pdfplumber==0.11.9
rouge-score==0.1.2
